feat(rewrite): update to 1.21.7

+++ common/src/main/kotlin/com/undefined/cassini/internal/info/PacketClickInformation.kt
@@ -2,4 +2,4 @@ package com.undefined.cassini.internal.info
 
 import org.bukkit.entity.Player
 
-data class PacketClickInformation(val player: Player, val slot: Int)
+data class PacketClickInformation(val player: Player, val slot: Short)
